Add support for client and peer certificates
This commit adds two additional command-line flags for specifying SSL
certificates to be used by drill when making HTTP requests.

`--cert` is used to specify a client certificate file (following the
`curl` option with the same name). Again similarly to curl, `--cacert`
is used to specify a PEM format certificate that should be used to
verify the remote peer.